services/service-a/requirements.txt
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
flask==3.0.0
requests==2.31.0